backend/requirements.txt
# backend/requirements.txt

praw
openai
firebase-admin
python-dotenv
